+++ app/src/test/scala/io/github/datacatering/datacaterer/core/util/UniqueFieldsUtilTest.scala
@@ -2,6 +2,7 @@ package io.github.datacatering.datacaterer.core.util
 
 import io.github.datacatering.datacaterer.api.model.Constants.{IS_PRIMARY_KEY, IS_UNIQUE}
 import io.github.datacatering.datacaterer.api.model.{Count, Field, PerFieldCount, Plan, Step, Task, TaskSummary}
+import org.apache.spark.util.sketch.BloomFilter
 
 import java.sql.Date
 
@@ -26,13 +27,15 @@ class UniqueFieldsUtilTest extends SparkSuite {
     val uniqueFields = uniqueFieldUtil.uniqueFieldsDf
     assertResult(2)(uniqueFields.size)
     assertResult(2)(uniqueFieldUtil.uniqueFieldsDf.size)
-    assert(uniqueFieldUtil.uniqueFieldsDf.head._2.isEmpty)
     val col = uniqueFields.filter(_._1.fields == List("account_id")).head
     assertResult("postgresAccount")(col._1.dataSource)
     assertResult("accounts")(col._1.step)
 
     val generatedData = sparkSession.createDataFrame(Seq(
-      Account("acc1", "peter"), Account("acc1", "john"), Account("acc2", "jack"), Account("acc3", "bob")
+      Account("acc1", "peter"),
+      Account("acc1", "john"),
+      Account("acc2", "jack"),
+      Account("acc3", "bob")
     ))
     val result = uniqueFieldUtil.getUniqueFieldsValues("postgresAccount.accounts", generatedData, Step())
 
@@ -41,27 +44,42 @@ class UniqueFieldsUtilTest extends SparkSuite {
     assertResult(3)(data.length)
     data.foreach(a => assert(expectedUniqueAccounts.contains(a)))
     assertResult(2)(uniqueFieldUtil.uniqueFieldsDf.size)
-    assertResult(3)(uniqueFieldUtil.uniqueFieldsDf.head._2.count())
-    val currentUniqueAcc = uniqueFieldUtil.uniqueFieldsDf.filter(_._1.fields == List("account_id")).head._2.collect().map(_.getString(0))
-    currentUniqueAcc.foreach(a => assert(expectedUniqueAccounts.contains(a)))
 
     val generatedData2 = sparkSession.createDataFrame(Seq(
       Account("acc1", "dog"), Account("acc3", "bob"), Account("acc4", "cat"), Account("acc5", "peter")
     ))
     val result2 = uniqueFieldUtil.getUniqueFieldsValues("postgresAccount.accounts", generatedData2, Step())
 
+    result2.cache()
+    result2.show()
     val data2 = result2.select("account_id", "name").collect()
-    val expectedUniqueNames = Array("peter", "jack", "bob", "cat")
-    val expectedUniqueAccounts2 = Array("acc1", "acc2", "acc3", "acc4")
 
     assertResult(1)(data2.length)
     assertResult("acc4")(data2.head.getString(0))
     assertResult("cat")(data2.head.getString(1))
+  }
 
-    val currentUniqueAcc2 = uniqueFieldUtil.uniqueFieldsDf.filter(_._1.fields == List("account_id")).head._2.collect().map(_.getString(0))
-    currentUniqueAcc2.foreach(a => assert(expectedUniqueAccounts2.contains(a)))
-    val currentUniqueName = uniqueFieldUtil.uniqueFieldsDf.filter(_._1.fields == List("name")).head._2.collect().map(_.getString(0))
-    currentUniqueName.foreach(a => assert(expectedUniqueNames.contains(a)))
+  test("Can filter out duplicate records from previous generated data") {
+    val generatedData = sparkSession.createDataFrame(Seq(
+      Account("acc1", "peter"),
+      Account("acc2", "john"),
+      Account("acc3", "jack"),
+      Account("acc4", "bob"),
+      Account("acc5", "alice")
+    ))
+
+    val bloomFilter = BloomFilter.create(1e7.toLong, 0.001) // 10M entries, 0.1% error rate
+    bloomFilter.putString("acc4|bob")
+    bloomFilter.putString("acc5|alice")
+    val bloomFilterBC = sparkSession.sparkContext.broadcast(bloomFilter)
+    val prevGen = UniqueFields("my-data-source", "step1", List("account_id", "name")) -> bloomFilterBC
+    val uniqueDF = new UniqueFieldsUtil(Plan(), List()).filterUniqueRecords(generatedData, prevGen)
+
+    val res = uniqueDF.collect()
+    assertResult(3)(res.length)
+    assert(res.exists(r => r.getString(0) == "acc1"))
+    assert(res.exists(r => r.getString(0) == "acc2"))
+    assert(res.exists(r => r.getString(0) == "acc3"))
   }
 
   test("Can identify the unique fields and create a data frame with unique values with per field count defined") {
@@ -93,10 +111,6 @@ class UniqueFieldsUtilTest extends SparkSuite {
     assertResult(3)(data.length)
     data.foreach(a => assert(expectedUniqueAccounts.contains(a)))
     assertResult(2)(uniqueFieldUtil.uniqueFieldsDf.size)
-    assertResult(3)(uniqueFieldUtil.uniqueFieldsDf.head._2.count())
-    assertResult(3)(uniqueFieldUtil.uniqueFieldsDf.last._2.count())
-    val currentUniqueAcc = uniqueFieldUtil.uniqueFieldsDf.filter(_._1.fields == List("account_id")).head._2.collect().map(_.getString(0))
-    currentUniqueAcc.foreach(a => assert(expectedUniqueAccounts.contains(a)))
   }
 
   test("Can identify the primary key fields, create a data frame with per field count defined, create unique values for the primary key fields and don't track globally unique values") {
@@ -137,9 +151,5 @@ class UniqueFieldsUtilTest extends SparkSuite {
     assertResult(5)(nameData.length)
     nameData.foreach(a => assert(expectedUniqueNames.contains(a)))
     assertResult(1)(uniqueFieldUtil.uniqueFieldsDf.size)
-    assertResult(0)(uniqueFieldUtil.uniqueFieldsDf.head._2.count())
-    assertResult(0)(uniqueFieldUtil.uniqueFieldsDf.last._2.count())
-    val trackingDf = uniqueFieldUtil.uniqueFieldsDf.filter(_._1.fields == List("account_id", "name")).head._2.collect()
-    assert(trackingDf.isEmpty)
   }
 }